refactor(route/iwara): extract shared Playwright and JSON helpers
- Add fetchJsonInPage helper for in-page JSON API calls
- Add getIwaraPage helper for shared Playwright page configuration
- Replace duplicated page.evaluate and getPlaywrightPage code across routes
