+++ lib/routes/iwara/utils.ts
@@ -1,6 +1,31 @@
+import { getPlaywrightPage, type Page } from '@/utils/playwright';
+
 export const rootUrl = 'https://www.iwara.tv';
 export const apiRootUrl = 'https://api.iwara.tv';
 export const imageRootUrl = 'https://i.iwara.tv';
+
+export const getIwaraPage = () =>
+    getPlaywrightPage(rootUrl, {
+        closeTimeout: 90 * 1000,
+        onBeforeLoad: async (page) => {
+            await page.route('**/*', (route) => {
+                const resourceType = route.request().resourceType();
+                ['document', 'script', 'xhr', 'fetch'].includes(resourceType) ? route.continue() : route.abort();
+            });
+        },
+        gotoConfig: {
+            waitUntil: 'domcontentloaded',
+        },
+    });
+
+export const fetchJsonInPage = (page: Page, url: string) =>
+    page.evaluate(async (u) => {
+        const res = await fetch(u);
+        if (!res.ok) {
+            throw new Error(`HTTP error! status: ${res.status}`);
+        }
+        return res.json();
+    }, url);
 
 export const typeMap = {
     video: 'Videos',
